Closes #2: Get CVV not CV_CODE from .env file
As described in #2, the README tells the user to create a `.env` file
with a `CVC` variable in it but the `bot.js` file references a `CV_CODE`
variable in the dotenv file which doesn't exists. It's actually called a
Card Verification Value (CVV) code and not a CVC code and hence the
variable names were adjusted to match that.